pyproject.toml
[project]
name = "vision-demo"
version = "0.6.0"
description = "End-to-end object detection demo with distributed training and inference"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.2",
    "torchvision>=0.17",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "minio>=7.2",
    "pycocotools>=2.0",
    "pillow>=10.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-multipart>=0.0.9",
    "aiohttp>=3.13.3",
    "tqdm>=4.67.3",
    "mlflow>=3.10.1",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
    "pre-commit>=3.7",
    "ruff>=0.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "PTH", "D"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov --cov-report=term-missing"

[tool.coverage.run]
source = ["src/vision_demo"]
branch = true

[tool.coverage.report]
fail_under = 0

[tool.semantic_release]
version_toml = ["pyproject.toml:project.version"]
allow_zero_version = true
major_on_zero = false

[tool.semantic_release.branches.main]
match = "main"

[tool.semantic_release.changelog]
mode = "update"
insertion_flag = "<!-- version list -->"

[tool.semantic_release.changelog.default_templates]
changelog_file = "CHANGELOG.md"
output_format = "md"

[tool.semantic_release.publish]
upload_to_vcs_release = true
